FIX: Explicitly convert data frame to float
Fixes warning: Arrays of bytes/strings is being converted
to decimal numbers if dtype='numeric'. This behavior is
deprecated in 0.24 and will be removed in 1.1 (renaming
of 0.26). Please convert your data to numeric values
explicitly instead.

+++ tests/test_pandas_inputs.py
@@ -34,7 +34,7 @@ def test_pandas_inputs(estimator_cls):
     X, y = load_whas500()
     X = X.iloc[:50]
     y = y[:50]
-    X = X.loc[:, ["age", "bmi", "chf", "gender"]]
+    X = X.loc[:, ["age", "bmi", "chf", "gender"]].astype(float)
 
     estimator = estimator_cls()
     if "kernel" in estimator.get_params():
